Add support for localization and timeline feature
The commits included the inclusion of the @angular/localize package for the support of localization in the application. Moreover, the timeline feature was also added to the AppComponent through the addition of the <app-timeline> element. The updates further include the registration of Spanish locale data, and the import of related UI modules.
This change was necessary to enhance the user-dynamicity of the application, and to include additional components that would aid in better application responsiveness and user interactivity.